Перенёс константы и функции, формирующие тестовые данные на их основе в /mock/trip-point.js
